Added Camera following
The camera now stays centered on the two players and zooms in and out. This way we can have a much larger stage. Min and Max zoom level can be changed so that it doesnt get too close or too far away. Weird snapping bug at beginning of game needs fixed still